+++ examples/oxfordflowers102/diffusion/train.go
@@ -15,8 +15,8 @@ import (
 	"github.com/gomlx/gomlx/ml/train/optimizers"
 	"github.com/gomlx/gomlx/types"
 	"github.com/gomlx/gomlx/types/shapes"
-	"github.com/gomlx/gomlx/types/slices"
 	"github.com/gomlx/gomlx/types/tensor"
+	"k8s.io/klog/v2"
 	"os"
 	"path"
 	"strings"
@@ -24,20 +24,31 @@ import (
 )
 
 var (
-	flagCheckpoint       = flag.String("checkpoint", "", "Directory save and load checkpoints from. If left empty, no checkpoints are created.")
-	flagCheckpointKeep   = flag.Int("checkpoint_keep", 20, "Number of checkpoints to keep, if --checkpoint is set.")
-	flagCheckpointPeriod = flag.Int("checkpoint_period", 60, "Period of time, in seconds, between checkpoints are saved.")
+	flagCheckpoint         = flag.String("checkpoint", "", "Directory save and load checkpoints from. If left empty, no checkpoints are created.")
+	flagCheckpointKeep     = flag.Int("checkpoint_keep", 20, "Number of checkpoints to keep, if --checkpoint is set.")
+	flagCheckpointPeriod   = flag.Int("checkpoint_period", 60, "Period of time, in seconds, between checkpoints are saved.")
+	flagCheckpointTakeMean = flag.Int("checkpoint_mean", 1, "If != 1, take the mean of the latest checkpoints. This is disabled (set to 1) if training.")
 )
 
 const (
 	NoiseSamplesFile       = "noise_samples.tensor"
+	FlowerIdsSamplesFile   = "flower_ids_samples.tensor"
 	GeneratedSamplesPrefix = "generated_samples_"
 )
 
 // LoadCheckpointToContext and attaches to it, so that it gets saved.
 //
-// It also loads the noise samples for this model.
-func LoadCheckpointToContext(ctx *context.Context) (checkpoint *checkpoints.Handler, noise tensor.Tensor) {
+// It also loads the noise (+flowerIds) samples for this model.
+// The idea is that at each evaluation checkpoint we generate the images for these fixed noise samples,
+// and one can observe the model quality evolving.
+//
+// For new models -- whose directory didn't previously exist, it does 2 things:
+//
+//   - It creates the noise + flowerIds samples used to monitor the model quality evolving.
+//   - It creates the file `args.txt` with a copy of the arguments used to create the model.
+//     Later, if the same model is used, it checks that the arguments match (with some exceptions),
+//     and warns about mismatches.
+func LoadCheckpointToContext(ctx *context.Context) (checkpoint *checkpoints.Handler, noise, flowerIds tensor.Tensor) {
 	Init()
 	if *flagCheckpoint == "" {
 		return
@@ -48,7 +59,9 @@ func LoadCheckpointToContext(ctx *context.Context) (checkpoint *checkpoints.Hand
 		checkpointPath = path.Join(DataDir, checkpointPath)
 	}
 	var err error
-	checkpoint, err = checkpoints.Build(ctx).Dir(checkpointPath).Keep(*flagCheckpointKeep).Done()
+	checkpoint, err = checkpoints.Build(ctx).Dir(checkpointPath).
+		Keep(*flagCheckpointKeep).TakeMean(*flagCheckpointTakeMean).
+		Done()
 	AssertNoError(err)
 
 	// Check if args file exist, if not create it.
@@ -77,19 +90,24 @@ func LoadCheckpointToContext(ctx *context.Context) (checkpoint *checkpoints.Hand
 		AssertNoError(err)
 	}
 
-	// Load/generate sampled noise.
-	noisePath := path.Join(checkpointPath, NoiseSamplesFile)
+	// Load/generate sampled noise/flowerIds.
+	noisePath, flowerIdsPath := path.Join(checkpointPath, NoiseSamplesFile), path.Join(checkpointPath, FlowerIdsSamplesFile)
 	noise, err = tensor.Load(noisePath)
 	if err == nil {
-		return
+		flowerIds, err = tensor.Load(flowerIdsPath)
+		if err == nil {
+			return
+		}
 	}
 	if !os.IsNotExist(err) {
 		AssertNoError(err)
 	}
 
-	// Create new noise -- and save it for future training.
+	// Create new noise and flower ids -- and save it for future training.
 	noise = GenerateNoise(*flagTrainGeneratedSamples)
+	flowerIds = GenerateFlowerIds(*flagTrainGeneratedSamples)
 	AssertNoError(noise.Local().Save(noisePath))
+	AssertNoError(flowerIds.Local().Save(flowerIdsPath))
 	return
 }
 
@@ -99,7 +117,7 @@ func isArgIrrelevant(arg string) bool {
 	if irrelevantArgs.Has(arg) {
 		return true
 	}
-	for _, prefix := range []string{"-steps", "--steps", "--batch", "--eval_batch"} {
+	for _, prefix := range []string{"-steps", "--steps", "--batch", "--eval_batch", "--checkpoint_mean", "--platform"} {
 		if strings.HasPrefix(arg, prefix) {
 			return true
 		}
@@ -108,15 +126,15 @@ func isArgIrrelevant(arg string) bool {
 }
 
 var (
-	flagNumSteps = flag.Int("steps", 2000, "Number of gradient descent steps to perform.")
-	flagPlots    = flag.Bool("plots", true, "Plots during training: perform periodic evaluations, "+
+	flagNumSteps = flag.Int("steps", 2000, "Number of gradient descent steps to perform in total "+
+		"-- this includes the steps already trained, if restarting training a model.")
+	flagPlots = flag.Bool("plots", true, "Plots during training: perform periodic evaluations, "+
 		"save results if --checkpoint is set and draw plots, if in a Jupyter notebook.")
 	flagKid = flag.Bool("kid", true, "If true, calculate Kernel Inception Distance (KID) on evaluation "+
 		"-- it is quite expensive.")
 
 	// Training hyperparameters:
-	flagOptimizer        = flag.String("optimizer", "adamw", fmt.Sprintf("Optimizer, options: %q", slices.SortedKeys(optimizers.KnownOptimizers)))
-	flagLearningRate     = flag.Float64("learning_rate", 0.0001, "Initial learning rate.")
+	flagLearningRate     = flag.Float64("learning_rate", 0.001, "Initial learning rate.")
 	flagL2Regularization = flag.Float64("l2_reg", 0, "L2 regularization on kernels. It doesn't interact well with --batch_norm.")
 	flagReport           = flag.Bool("report", true, "If true generate evaluation report at end of training.")
 	flagRngReset         = flag.Bool("rng_reset", true, "If true will reset the random number generator state with a new random value -- useful when continuing training.")
@@ -129,6 +147,7 @@ var (
 
 func TrainModel() {
 	Init()
+	*flagCheckpointTakeMean = 1 // Disable mean of checkpoints if training.
 	trainDS, validationDS := CreateInMemoryDatasets()
 	trainEvalDS := trainDS.Copy()
 
@@ -142,13 +161,16 @@ func TrainModel() {
 	ctx.SetParam(layers.L2RegularizationKey, *flagL2Regularization)
 
 	// Checkpoints saving.
-	checkpoint, noise := LoadCheckpointToContext(ctx)
+	checkpoint, noise, flowerIds := LoadCheckpointToContext(ctx)
 	if *flagRngReset {
 		ctx.RngStateReset()
 	}
-	globalStep := optimizers.GetGlobalStepVar(ctx).Value().Value()
+	globalStep := optimizers.GetGlobalStepVar(ctx).Value().Value().(int)
 	if globalStep != 0 {
 		fmt.Printf("Restarting training from global_step=%d\n", globalStep)
+	}
+	if globalStep >= *flagNumSteps {
+		klog.Exitf("Current global step %d >= target --steps=%d, exiting.", globalStep, *flagNumSteps)
 	}
 
 	// Custom loss: model returns scalar loss as the second element of the predictions.
@@ -172,7 +194,11 @@ func TrainModel() {
 		optimizers.Adam().WeightDecay(1e-4).Done(),
 		[]metrics.Interface{movingImagesLoss}, // trainMetrics
 		[]metrics.Interface{meanImagesLoss})   // evalMetrics
-
+	if *flagNanLogger {
+		trainer.OnExecCreation(func(exec *context.Exec, _ train.GraphType) {
+			nanLogger.Attach(exec)
+		})
+	}
 	// Use standard training loop.
 	loop := train.NewLoop(trainer)
 	loop.ReadGlobalStep(ctx)
@@ -181,7 +207,11 @@ func TrainModel() {
 	// Attach a checkpoint.
 	if checkpoint != nil {
 		period := time.Second * time.Duration(*flagCheckpointPeriod)
-		train.PeriodicCallback(loop, period, true, "saving checkpoint", 100, checkpoint.OnStepFn)
+		train.PeriodicCallback(loop, period, true, "saving checkpoint", 100,
+			func(loop *train.Loop, metrics []tensor.Tensor) error {
+				fmt.Printf("\n[saving checkpoint@%d] [median train step (ms): %d]\n", loop.LoopStep, loop.MedianTrainStepDuration().Milliseconds())
+				return checkpoint.Save()
+			})
 	}
 
 	// Monitoring training: plots, generator of images, kid evaluator.
@@ -195,7 +225,7 @@ func TrainModel() {
 		plots.DynamicUpdates()
 	}
 
-	generator := NewImagesGenerator(ctx, noise, 20, 0)
+	generator := NewImagesGenerator(ctx, noise, flowerIds, 20, 0)
 
 	var kid *KidGenerator
 	if *flagKid {
@@ -210,7 +240,7 @@ func TrainModel() {
 		})
 
 	// Loop for given number of steps.
-	_, err := loop.RunSteps(trainDS, *flagNumSteps)
+	_, err := loop.RunSteps(trainDS, *flagNumSteps-globalStep)
 	if err != nil {
 		fmt.Printf("\nFailed: %v\n\n", err)
 		return
@@ -236,6 +266,7 @@ func TrainModel() {
 func TrainingMonitor(checkpoint *checkpoints.Handler, loop *train.Loop, metrics []tensor.Tensor,
 	plots *margaid.Plots, generator *ImagesGenerator, kid *KidGenerator) error {
 
+	fmt.Printf("\n[... evaluating@%d ...] [median train step (ms): %d]\n", loop.LoopStep, loop.MedianTrainStepDuration().Milliseconds())
 	// Save checkpoint, just in case.
 	if checkpoint == nil {
 		// Only works if there is a model directory.
@@ -265,7 +296,7 @@ func TrainingMonitor(checkpoint *checkpoints.Handler, loop *train.Loop, metrics 
 func DisplayTrainingPlots() {
 	Init()
 	ctx := context.NewContext(manager)
-	checkpoint, _ := LoadCheckpointToContext(ctx)
+	checkpoint, _, _ := LoadCheckpointToContext(ctx)
 	if checkpoint == nil {
 		fmt.Printf("You must set --checkpoint='model_sub_dir'!")
 		return