+++ docs/CHANGELOG.md
@@ -1,5 +1,17 @@
 # GoMLX changelog
 
+## v0.4.1
+
+* Diffusion example: Added conditioning on flower type; Improved documentation; several other small improvements.
+* NanLogger: added tool to report back (with stack trace and scope) on the occurrences of NaN/Inf in the computation
+  graph.
+* Checkpoints: added `Handler.LoadedVariables()` method for inspection of loaded checkpoint. 
+* Bug fixes: 
+  * RandomNormal: fixed rare numerical issues in RandomNormal, that would generate -Inf.
+  * Context: some rare condition on feeding variable values to executor.
+  * InMemory dataset: handling cases where dataset returns the same tensor as input and label.
+* Slices: refactored `IotaSlice()` to `Iota[T number]()`.
+
 ## v0.4.0
 
 * Models: Diffusion example model (working draft); added Kernel Inception Distance (KID) metric implementation.